+++ kernel/src/main.c
@@ -3,8 +3,8 @@
 
 uint64_t kmain(bootinfo *bootinfo){
 	KMESSAGE("ARCH %s", ARCH_STRING);
-	KMESSAGE("Framebuffer %dx%d at 0x%x (%d MB)", fb->width, fb->height, fb->buffer, BYTES_TO_MB(fb->size));
-	KMESSAGE("Available system memory %d MB", BYTES_TO_MB(heap.free_memory));
+	KMESSAGE("Framebuffer %dx%d at 0x%x (%d MB)", (uint64_t)fb->width, (uint64_t)fb->height, fb->buffer, (uint64_t)BYTES_TO_MB(fb->size));
+	KMESSAGE("Available system memory %d MB", (uint64_t)BYTES_TO_MB(heap.free_memory));
 	kprint("Welcome to l00n1x !\n");
 	//while(1){
 		task_new(
